+++ synth/src/osc.rs
@@ -10,7 +10,8 @@ pub struct Osc {
 
 impl Signal for Osc {
     fn next_value(&mut self) -> f32 {
-        self.clock.time += self.frequency.next_value() / self.clock.sample_rate;
+        self.clock.time +=
+            self.frequency.next_value() / self.clock.sample_rate as f32;
         self.clock.time %= 1.;
 
         self.offset + (self.wave)(self.clock.time) * self.amplitude
